requirements.txt.txt
streamlit
txtai
PyPDF2
deep-translator
langdetect
nlpaug
Pillow
pytesseract
python-docx
